requirements.txt
streamlit
requests
transformers
python-dotenv
PyMuPDF
chromadb
sentence-transformers
fastapi
uvicorn